chore(ui): clean up the type declaration file
