+++ openmetadata-ui/src/main/resources/ui/src/components/AddService/Steps/ConfigureService.tsx
@@ -11,11 +11,11 @@
  *  limitations under the License.
  */
 
-import { EditorContentRef } from 'Models';
 import React, { useRef } from 'react';
 import { errorMsg, requiredField } from '../../../utils/CommonUtils';
 import { Button } from '../../buttons/Button/Button';
 import RichTextEditor from '../../common/rich-text-editor/RichTextEditor';
+import { EditorContentRef } from '../../common/rich-text-editor/RichTextEditor.interface';
 import { Field } from '../../Field/Field';
 import { ConfigureServiceProps } from './Steps.interface';
