+++ openmetadata-ui/src/main/resources/ui/src/components/EntityTable/EntityTable.interface.ts
@@ -11,10 +11,10 @@
  *  limitations under the License.
  */
 
-import { EntityFieldThreads } from 'Models';
 import { ReactNode } from 'react';
 import { ThreadType } from '../../generated/api/feed/createThread';
 import { Column, ColumnJoins, Table } from '../../generated/entity/data/table';
+import { EntityFieldThreads } from '../../interface/feed.interface';
 
 export interface EntityTableProps {
   tableColumns: Column[];
