+++ openmetadata-ui/src/main/resources/ui/src/components/TopicDetails/TopicDetails.test.tsx
@@ -17,12 +17,15 @@ import {
   render,
   screen,
 } from '@testing-library/react';
-import { LeafNodes, LoadingNodeState } from 'Models';
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
 import { EntityReference } from '../../generated/type/entityReference';
 import { Paging } from '../../generated/type/paging';
 import { TagLabel } from '../../generated/type/tagLabel';
+import {
+  LeafNodes,
+  LoadingNodeState,
+} from '../EntityLineage/EntityLineage.interface';
 import TopicDetails from './TopicDetails.component';
 import { TOPIC_DETAILS } from './TopicDetails.mock';
