+++ openmetadata-ui/src/main/resources/ui/src/components/common/OwnerWidget/OwnerWidgetWrapper.component.tsx
@@ -13,7 +13,7 @@
 
 import { AxiosError } from 'axios';
 import { debounce, isEqual, lowerCase } from 'lodash';
-import { Status } from 'Models';
+import { LoadingState } from 'Models';
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { default as AppState, default as appState } from '../../../AppState';
 import { getGroupTypeTeams } from '../../../axiosAPIs/userAPI';
@@ -54,7 +54,7 @@ const OwnerWidgetWrapper = ({
 }: OwnerWidgetWrapperProps) => {
   const { isAuthDisabled } = useAuthContext();
   const { isAdminUser } = useAuth();
-  const [statusOwner, setStatusOwner] = useState<Status>('initial');
+  const [statusOwner, setStatusOwner] = useState<LoadingState>('initial');
 
   const [listOwners, setListOwners] = useState<
     {
