+++ openmetadata-ui/src/main/resources/ui/src/components/common/description/DescriptionV1.tsx
@@ -15,11 +15,11 @@ import { Space, Tooltip, Typography } from 'antd';
 import classNames from 'classnames';
 import { t } from 'i18next';
 import { isUndefined } from 'lodash';
-import { EntityFieldThreads } from 'Models';
 import React, { Fragment } from 'react';
 import { EntityField } from '../../../constants/Feeds.constants';
 import { NO_PERMISSION_FOR_ACTION } from '../../../constants/HelperTextUtil';
 import { Table } from '../../../generated/entity/data/table';
+import { EntityFieldThreads } from '../../../interface/feed.interface';
 import { getEntityFeedLink } from '../../../utils/EntityUtils';
 import SVGIcons, { Icons } from '../../../utils/SvgUtils';
 import { ModalWithMarkdownEditor } from '../../Modals/ModalWithMarkdownEditor/ModalWithMarkdownEditor';
