+++ openmetadata-ui/src/main/resources/ui/src/pages/TasksPage/RequestDescriptionPage/RequestDescriptionPage.tsx
@@ -16,7 +16,7 @@ import { useForm } from 'antd/lib/form/Form';
 import { AxiosError } from 'axios';
 import { capitalize, isNil } from 'lodash';
 import { observer } from 'mobx-react';
-import { EditorContentRef, EntityTags } from 'Models';
+import { EntityTags } from 'Models';
 import React, {
   useCallback,
   useEffect,
@@ -29,6 +29,7 @@ import AppState from '../../../AppState';
 import { postThread } from '../../../axiosAPIs/feedsAPI';
 import ProfilePicture from '../../../components/common/ProfilePicture/ProfilePicture';
 import RichTextEditor from '../../../components/common/rich-text-editor/RichTextEditor';
+import { EditorContentRef } from '../../../components/common/rich-text-editor/RichTextEditor.interface';
 import TitleBreadcrumb from '../../../components/common/title-breadcrumb/title-breadcrumb.component';
 import { FQN_SEPARATOR_CHAR } from '../../../constants/char.constants';
 import { EntityField } from '../../../constants/Feeds.constants';
