+++ openmetadata-ui/src/main/resources/ui/src/pages/TasksPage/UpdateDescriptionPage/UpdateDescriptionPage.tsx
@@ -15,7 +15,7 @@ import { Button, Card, Form, FormProps, Input, Space } from 'antd';
 import { useForm } from 'antd/lib/form/Form';
 import { AxiosError } from 'axios';
 import { capitalize, isEmpty, isNil, isUndefined } from 'lodash';
-import { EditorContentRef, EntityTags } from 'Models';
+import { EntityTags } from 'Models';
 import React, {
   useCallback,
   useEffect,
@@ -27,6 +27,7 @@ import { useHistory, useLocation, useParams } from 'react-router-dom';
 import AppState from '../../../AppState';
 import { postThread } from '../../../axiosAPIs/feedsAPI';
 import ProfilePicture from '../../../components/common/ProfilePicture/ProfilePicture';
+import { EditorContentRef } from '../../../components/common/rich-text-editor/RichTextEditor.interface';
 import TitleBreadcrumb from '../../../components/common/title-breadcrumb/title-breadcrumb.component';
 import { FQN_SEPARATOR_CHAR } from '../../../constants/char.constants';
 import { EntityField } from '../../../constants/Feeds.constants';
