+++ openmetadata-ui/src/main/resources/ui/src/pages/TasksPage/shared/DescriptionTabs.tsx
@@ -14,9 +14,9 @@
 import { Tabs } from 'antd';
 import { Change } from 'diff';
 import { isEqual } from 'lodash';
-import { EditorContentRef } from 'Models';
 import React, { useState } from 'react';
 import RichTextEditor from '../../../components/common/rich-text-editor/RichTextEditor';
+import { EditorContentRef } from '../../../components/common/rich-text-editor/RichTextEditor.interface';
 import RichTextEditorPreviewer from '../../../components/common/rich-text-editor/RichTextEditorPreviewer';
 import { getDescriptionDiff } from '../../../utils/TasksUtils';
 import { DiffView } from './DiffView';
