+++ openmetadata-ui/src/main/resources/ui/src/pages/teams/AddTeamForm.tsx
@@ -14,11 +14,11 @@
 import { Form, Input, Modal, Select } from 'antd';
 import { AxiosError } from 'axios';
 import { isUndefined, toLower, trim } from 'lodash';
-import { EditorContentRef } from 'Models';
 import React, { useEffect, useMemo, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { getTeams } from '../../axiosAPIs/teamsAPI';
 import RichTextEditor from '../../components/common/rich-text-editor/RichTextEditor';
+import { EditorContentRef } from '../../components/common/rich-text-editor/RichTextEditor.interface';
 import { Team, TeamType } from '../../generated/entity/teams/team';
 import jsonData from '../../jsons/en';
 import { isUrlFriendlyName } from '../../utils/CommonUtils';
