+++ openmetadata-ui/src/main/resources/ui/src/pages/tour-page/TourPage.component.tsx
@@ -13,11 +13,11 @@
 
 import { noop } from 'lodash';
 import { observer } from 'mobx-react';
-import { LeafNodes } from 'Models';
 import React, { useEffect, useState } from 'react';
 import { useLocation } from 'react-router-dom';
 import AppState from '../../AppState';
 import DatasetDetails from '../../components/DatasetDetails/DatasetDetails.component';
+import { LeafNodes } from '../../components/EntityLineage/EntityLineage.interface';
 import Explore from '../../components/Explore/Explore.component';
 import MyData from '../../components/MyData/MyData.component';
 import { MyDataProps } from '../../components/MyData/MyData.interface';
