+++ openmetadata-ui/src/main/resources/ui/src/utils/FeedElementUtils.tsx
@@ -12,11 +12,11 @@
  */
 
 import { isEmpty, isEqual, isUndefined } from 'lodash';
-import { EntityFieldThreads } from 'Models';
 import React, { Fragment } from 'react';
 import { entityUrlMap } from '../constants/Feeds.constants';
 import { ThreadType } from '../generated/entity/feed/thread';
 import { EntityReference } from '../generated/entity/teams/user';
+import { EntityFieldThreads } from '../interface/feed.interface';
 import { getEntityFeedLink } from './EntityUtils';
 import { getThreadField } from './FeedUtils';
 import SVGIcons, { Icons } from './SvgUtils';
