+++ openmetadata-ui/src/main/resources/ui/src/utils/FeedUtils.tsx
@@ -17,11 +17,7 @@ import { AxiosError } from 'axios';
 import { Operation } from 'fast-json-patch';
 import i18next from 'i18next';
 import { isEqual } from 'lodash';
-import {
-  EntityFieldThreadCount,
-  EntityFieldThreads,
-  EntityThreadField,
-} from 'Models';
+
 import React from 'react';
 import Showdown from 'showdown';
 import TurndownService from 'turndown';
@@ -52,6 +48,11 @@ import {
 import { EntityType, FqnPart, TabSpecificField } from '../enums/entity.enum';
 import { SearchIndex } from '../enums/search.enum';
 import { Thread, ThreadType } from '../generated/entity/feed/thread';
+import {
+  EntityFieldThreadCount,
+  EntityFieldThreads,
+  EntityThreadField,
+} from '../interface/feed.interface';
 import jsonData from '../jsons/en';
 import {
   getEntityPlaceHolder,
